error handler receives Curly::Response as well as stringified error
drive-by:
  * fix a bug where a 2xx response > 200 would have triggered error
    handling
  * explicitly return response.body in success cases

Additionally: previously, if a SwigClient were configured without
an error_handler, and it received a non-200 response, then the
response.body would have been returned. Now if a SwigClient is configured
without an error_handler and it receives a non-2xx response, then
we don't fall through to returning response.body. The previous behavior
looked like a bug to me, which is why I changed this as well.